types and resolvers seperated out into folders.  Renamed schema folder to types.  Created another folder for resolvers (queries and mutations).  Associations between post and user, or post and reply for example means we will have cross pollination between types.  Therefore I have added merge-graphql-schemas to first merge types and resolvers seperately, and THEN merge the types and resolvers in src/controllers/graphql
