+++ package.json
@@ -50,6 +50,7 @@
     "koa-static": "^4.0.2",
     "lightrpc": "^0.0.5",
     "lodash": "^4.17.10",
+    "merge-graphql-schemas": "^1.5.3",
     "mysql": "^2.15.0",
     "mysql2": "^1.5.3",
     "node-schedule": "^1.3.0",
