+++ src/controllers/graphql/index.js
@@ -1,0 +1,6 @@
+const { makeExecutableSchema } = require('graphql-tools');
+
+const typeDefs = require('./types');
+const resolvers = require('./resolvers');
+
+module.exports = makeExecutableSchema({ typeDefs, resolvers });
