+++ src/controllers/graphql/resolvers/index.js
@@ -1,0 +1,10 @@
+const { mergeResolvers } = require('merge-graphql-schemas');
+const userResolver = require('./user');
+const postResolver = require('./post');
+
+const resolvers = [
+  userResolver,
+  postResolver,
+];
+
+module.exports = mergeResolvers(resolvers);
