+++ src/controllers/graphql/resolvers/post/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./postResolver');
