+++ src/controllers/graphql/resolvers/user/userResolver.js
@@ -1,0 +1,16 @@
+const db = require('../../connectors');
+
+const USer = db.sequelize.models.user;
+
+module.exports = {
+  Query: {
+    user(_, args) {
+      return User.find({ where: args });
+    }
+  },
+  User: {
+    posts(author) {
+      return author.getPosts();
+    }
+  }
+};
