--- src/controllers/graphql/schemas/comment/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./comment');
