--- src/controllers/graphql/schemas/post/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./post');
