--- src/controllers/graphql/schemas/user/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./user');
