+++ src/controllers/graphql/types/index.js
@@ -1,0 +1,10 @@
+const { mergeTypes } = require('merge-graphql-schemas');
+const postType = require('./post');
+const userType = require('./user');
+
+const types = [
+  postType,
+  userType,
+];
+
+module.exports = mergeTypes(types);
